Revert "PID trying 1."
This reverts commit 2029327fa8027f0f3190c9eac784df524eecc969.

+++ PID.h
@@ -13,10 +13,10 @@ typedef struct _pid pid;
 struct _pid{
     //unsigned int Setfreq;            //定义设定值
    // unsigned int Actualfreq;         //定义实际值
-    int err;                 //定义偏差值
-    int err_last;            //定义上一个偏差值
+    float err;                 //定义偏差值
+    float err_last;            //定义上一个偏差值
     float Kp,Ki,Kd;            //定义比例、积分、微分系数
-    int output;             //定义电压值（控制执行器的变量）
+    float output;             //定义电压值（控制执行器的变量）
     float integral;            //定义积分值
 };
